requirements.txt
ansi2html==1.8.0
azure-core==1.29.5
azure-identity==1.15.0
azure-storage-blob==12.19.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.3.0
cryptography==41.0.7
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
flask==3.0.0
idna==3.6
importlib-metadata==7.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.26.0
msal-extensions==1.1.0
nest-asyncio==1.5.8
numpy==1.24.4
packaging==23.2
pandas==2.0.3
plotly==5.18.0
portalocker==2.8.2
py4j==0.10.9.7
pycparser==2.21
PyJWT==2.8.0
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing-extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0